+++ receiver/hostmetricsreceiver/internal/scraper/pagingscraper/paging_scraper_windows.go
@@ -42,8 +42,8 @@ const (
 
 // scraper for Paging Metrics
 type scraper struct {
-	config    *Config
-	startTime pdata.Timestamp
+	config *Config
+	mb     *metadata.MetricsBuilder
 
 	perfCounterScraper perfcounters.PerfCounterScraper
 
@@ -63,7 +63,7 @@ func (s *scraper) start(context.Context, component.Host) error {
 		return err
 	}
 
-	s.startTime = pdata.Timestamp(bootTime * 1e9)
+	s.mb = metadata.NewMetricsBuilder(s.config.Metrics, metadata.WithStartTime(pdata.Timestamp(bootTime*1e9)))
 
 	return s.perfCounterScraper.Initialize(memory)
 }
@@ -74,12 +74,12 @@ func (s *scraper) scrape(context.Context) (pdata.Metrics, error) {
 
 	var errors scrapererror.ScrapeErrors
 
-	err := s.scrapeAndAppendPagingUsageMetric(metrics)
+	err := s.scrapePagingUsageMetric(metrics)
 	if err != nil {
 		errors.AddPartial(pagingUsageMetricsLen, err)
 	}
 
-	err = s.scrapeAndAppendPagingOperationsMetric(metrics)
+	err = s.scrapePagingOperationsMetric(metrics)
 	if err != nil {
 		errors.AddPartial(pagingMetricsLen, err)
 	}
@@ -87,40 +87,26 @@ func (s *scraper) scrape(context.Context) (pdata.Metrics, error) {
 	return md, errors.Combine()
 }
 
-func (s *scraper) scrapeAndAppendPagingUsageMetric(metrics pdata.MetricSlice) error {
+func (s *scraper) scrapePagingUsageMetric(metrics pdata.MetricSlice) error {
 	now := pdata.NewTimestampFromTime(time.Now())
 	pageFiles, err := s.pageFileStats()
 	if err != nil {
 		return err
 	}
 
-	idx := metrics.Len()
-	metrics.EnsureCapacity(idx + pagingUsageMetricsLen)
-	s.initializePagingUsageMetric(metrics.AppendEmpty(), now, pageFiles)
+	s.recordPagingUsageDataPoints(now, pageFiles)
+	s.mb.Emit(metrics)
 	return nil
 }
 
-func (s *scraper) initializePagingUsageMetric(metric pdata.Metric, now pdata.Timestamp, pageFiles []*pageFileStats) {
-	metadata.Metrics.SystemPagingUsage.Init(metric)
-
-	idps := metric.Sum().DataPoints()
-	idps.EnsureCapacity(2 * len(pageFiles))
-
+func (s *scraper) recordPagingUsageDataPoints(now pdata.Timestamp, pageFiles []*pageFileStats) {
 	for _, pageFile := range pageFiles {
-		initializePagingUsageDataPoint(idps.AppendEmpty(), now, pageFile.deviceName, metadata.AttributeState.Used, int64(pageFile.usedBytes))
-		initializePagingUsageDataPoint(idps.AppendEmpty(), now, pageFile.deviceName, metadata.AttributeState.Free, int64((pageFile.freeBytes)))
+		s.mb.RecordSystemPagingUsageDataPoint(now, int64(pageFile.usedBytes), pageFile.deviceName, metadata.AttributeState.Used)
+		s.mb.RecordSystemPagingUsageDataPoint(now, int64(pageFile.freeBytes), pageFile.deviceName, metadata.AttributeState.Free)
 	}
 }
 
-func initializePagingUsageDataPoint(dataPoint pdata.NumberDataPoint, now pdata.Timestamp, deviceLabel string, stateLabel string, value int64) {
-	attributes := dataPoint.Attributes()
-	attributes.InsertString(metadata.Attributes.Device, deviceLabel)
-	attributes.InsertString(metadata.Attributes.State, stateLabel)
-	dataPoint.SetTimestamp(now)
-	dataPoint.SetIntVal(value)
-}
-
-func (s *scraper) scrapeAndAppendPagingOperationsMetric(metrics pdata.MetricSlice) error {
+func (s *scraper) scrapePagingOperationsMetric(metrics pdata.MetricSlice) error {
 	now := pdata.NewTimestampFromTime(time.Now())
 
 	counters, err := s.perfCounterScraper.Scrape()
@@ -141,26 +127,13 @@ func (s *scraper) scrapeAndAppendPagingOperationsMetric(metrics pdata.MetricSlic
 	if len(memoryCounterValues) > 0 {
 		idx := metrics.Len()
 		metrics.EnsureCapacity(idx + pagingMetricsLen)
-		initializePagingOperationsMetric(metrics.AppendEmpty(), s.startTime, now, memoryCounterValues[0])
+		s.recordPagingOperationsDataPoints(now, memoryCounterValues[0])
 	}
-
+	s.mb.Emit(metrics)
 	return nil
 }
 
-func initializePagingOperationsMetric(metric pdata.Metric, startTime, now pdata.Timestamp, memoryCounterValues *perfcounters.CounterValues) {
-	metadata.Metrics.SystemPagingOperations.Init(metric)
-
-	idps := metric.Sum().DataPoints()
-	idps.EnsureCapacity(2)
-	initializePagingOperationsDataPoint(idps.AppendEmpty(), startTime, now, metadata.AttributeDirection.PageIn, memoryCounterValues.Values[pageReadsPerSec])
-	initializePagingOperationsDataPoint(idps.AppendEmpty(), startTime, now, metadata.AttributeDirection.PageOut, memoryCounterValues.Values[pageWritesPerSec])
-}
-
-func initializePagingOperationsDataPoint(dataPoint pdata.NumberDataPoint, startTime, now pdata.Timestamp, directionLabel string, value int64) {
-	attributes := dataPoint.Attributes()
-	attributes.InsertString(metadata.Attributes.Type, metadata.AttributeType.Major)
-	attributes.InsertString(metadata.Attributes.Direction, directionLabel)
-	dataPoint.SetStartTimestamp(startTime)
-	dataPoint.SetTimestamp(now)
-	dataPoint.SetIntVal(value)
+func (s *scraper) recordPagingOperationsDataPoints(now pdata.Timestamp, memoryCounterValues *perfcounters.CounterValues) {
+	s.mb.RecordSystemPagingFaultsDataPoint(now, memoryCounterValues.Values[pageReadsPerSec], metadata.AttributeDirection.PageIn)
+	s.mb.RecordSystemPagingFaultsDataPoint(now, memoryCounterValues.Values[pageWritesPerSec], metadata.AttributeDirection.PageOut)
 }